# api/api-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Express API with Swagger",
        "version": "1.0.0",
        "description": "REST API documentation using Swagger/OpenAPI"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "summary": "Returns all orders",
                "tags": [
                    "Orders"
                ],
                "responses": {
                    "200": {
                        "description": "List of all orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new order",
                "tags": [
                    "Orders"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "summary": "Get an order by ID",
                "tags": [
                    "Orders"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Order ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "put": {
                "summary": "Update an order",
                "tags": [
                    "Orders"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Order ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete an order",
                "tags": [
                    "Orders"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Order ID"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Order deleted successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
...
        "/api/purchase-orders/{id}": {
            "get": {
                "summary": "Get a purchase order by ID",
                "tags": [
                    "PurchaseOrders"
                ],
                "responses": {
                    "200": {
                        "description": "Purchase order found"
                    },
                    "404": {
                        "description": "Purchase order not found"
                    }
                }
            }
        },
        "/api/purchase-orders/{id}/submit": {
            "post": {
                "summary": "Submit a draft purchase order",
                "tags": [
                    "PurchaseOrders"
                ],
                "responses": {
                    "200": {
                        "description": "Purchase order submitted"
                    },
                    "409": {
                        "description": "Invalid status transition"
                    },
                    "502": {
                        "description": "Supplier notification dispatch failed"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Order": {
                "type": "object",
                "required": [
                    "orderId",
                    "branchId",
                    "orderDate"
                ],
                "properties": {
                    "orderId": {
                        "type": "integer",
                        "description": "The unique identifier for the order"
                    },
                    "branchId": {
                        "type": "integer",
                        "description": "The ID of the branch that placed the order"
                    },
                    "orderDate": {
                        "type": "string",
                        "format": "date-time",
                        "description": "The date and time when the order was placed"
                    },
                    "status": {
                        "type": "string",
                        "description": "The current status of the order",
                        "enum": [
                            "pending",
                            "processing",
                            "shipped",
                            "delivered",
                            "cancelled"
                        ]
                    },
                    "totalAmount": {
                        "type": "number",
                        "format": "float",
                        "description": "The total amount of the order"
                    }
                }
            },
...
        }
    },
...
}